+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * Reuse `<EndOfItem>` component from `stripes-data-transfer-components` repository (UIDATIMP-571)
 * Fix field mapping for Item record check in/check out note (UIDATIMP-554)
 * Field mappings: Repeatable field dropdown Validation (UIDATIMP-508)
+* Inventory field mapping: Instance, Holding, Item: add REMOVE option (UIDATIMP-567)
+* Field Mapping Profile details: MARC Bib from MARC Bib 10 - View details screen (UIDATIMP-494)
 
 ### Bugs fixed:
 * Fix rendering qualifier sections with old data in match profiles details (UIDATIMP-481)